First successful test run!!!

+++ PrimeNum/PrimeNumbersGenerator.hpp
@@ -4,7 +4,9 @@
 //
 //  Created by Mick Maciel on 27/02/2024.
 //
+
 #include "main.hpp"
+#include "PrimeNumbersTable.hpp"
 
 #ifndef PrimeNumbersGenerator_hpp
 #define PrimeNumbersGenerator_hpp
@@ -18,7 +20,7 @@ R
 class PrimeNumbersGenerator {
 public:
   bool isPrime(int number) const;
-  void generate(PrimeNumbersTable *primeTable, uint24 inputLimit);
+  void generate(PrimeNumbersTable *primeTable, unsigned int inputLimit);
 };
 
 #endif /* PrimeNumbersGenerator_hpp */
